==> 4_results_no_comments.Rmd <==
---
title: "Results for the census of statisticians on ethics committees"
author: "Adrian Barnett"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output: word_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message=FALSE, error=FALSE, comment='', dpi=400)
options(width=1000) # Wide pages
options(scipen=999) # avoid scientific presentation
#library(binom) # for confidence intervals
library(tidyr)
library(stringr)
library(dplyr)
options(dplyr.summarise.inform = FALSE)
library(janitor) # for tabyl
library(flextable)
library(xtable) # for latex
library(EnvStats) # for triangular distribution
source('99_functions.R')
TeachingDemos::char2seed('Lincoln')

# graphics set up
library(ggplot2)
library(visdat) # for item missing plots
library(gridExtra)
g.theme = theme_bw() + theme(panel.grid.minor = element_blank())

# from 3_blind_comments_do_not_share.R
load('data/3_HREC_Responses.RData')
n_non_consent = sum(data$q1_1 != 'I consent to participate')
# only those who consented
data = filter(data, q1_1 == 'I consent to participate') %>%
  select(-q1_1)

# questionnaires sent, from 1_send_emails.R
load('data/1_number_sent.RData')
# from 0_read_registered_hrec.R
load('data/0_hrecs.RData')
```

# Broken and incorrect links

There were many hyper links in the sampling frame that were broken or did not resolve to HREC's web page.

```{r}
tab = tabyl(hrecs, nhmrc_link_resolve) %>%
  mutate(percent = round(percent*100))
names(tab)[1] = 'Link resolved'
ftab = flextable(tab) %>%
  theme_box() %>%
  colformat_double(j=3, digits=0) %>%
  autofit()
ftab
```



# CONSORT flow chart

```{r}
source('4_consort_flow.R')
knitr::include_graphics(path='figures/4_consort_flow.jpg')
```

The diagram shows the progress from the sampling frame to the number analysed. 

# Recruitment

Plot of recruitment over time.

```{r}
source('4_plot_recruitment_over_time.R')
print(gplot)
# for text
first = min(data$recorded_date)
last = max(data$recorded_date)
days_recruit = as.numeric(last - first) + 1
first = format(first, '%d %b %Y')
last = format(last, '%d %b %Y')
```

The total number recruited is `r nrow(data)`. The first participant was recruited on `r first` and the last on `r last`, which is `r days_recruit` days.

There were `r n_non_consent` respondents who did not consent to participate according to the yes/no consent question in the form.

The total number of questionnaires sent was `r n_sent` so the response rate is `r round(100*nrow(data)/n_sent)`%.


# Meta-data on questions

## Time taken to answer questions

```{r}
stats = summarise(data,
            Q10 = round(quantile(duration_mins, 0.05)),
            Q25 = round(quantile(duration_mins, 0.25)),
            median = round(median(duration_mins)),
            Q75 = round(quantile(duration_mins, 0.75)),
            Q90 = round(quantile(duration_mins, 0.95)))
#
ftab = flextable(stats) %>%
  theme_box() %>%
  autofit()
ftab
```

These results are only for those who answered some of the questions and exclude the "click through" respondents. 
The summary statistics show the time in minutes to complete the online questions. Q[x] is the *x*th percentile.

The very long times are almost certainly due to respondents leaving the browser window open and returning to complete the questionnaire. This is possibly because they had to gather further information. 

## Questionnaire progress as a percent

The table below shows the questionnaire progress as a percent, with the results grouped into four categories.

```{r}
#
tab = tabyl(data, progress_cat) %>%
  mutate(percent = round(percent*100))
names(tab)[1] = 'Progress %'
ftab = flextable(tab) %>%
  theme_box() %>%
  colformat_double(j=3, digits=0) %>%
  autofit()
ftab
# for text
finished = sum(data$progress==100)
finished_p = round(100*finished / nrow(data))
```

Most respondents completed most of the questions. 

# Questions

### What is your current role(s) on the committee?

Respondents could tick multiple boxes.

```{r current_role}
tab = make_tab_tick_all(indata = data, var_start = 'q3') %>%
  rename('Role' = 'name') %>%
  mutate(Role = nice_rename(Role))
ftab = flextable(tab) %>%
  theme_box() %>%
#  colformat_double(j=3, digits=0) %>%
  autofit()
ftab
```

The percent for "No answer" is for all participants who could have answered the question. The other percents use a denominator of respondents who gave at least one answer.


### How many full members are on the current committee including the chair?

```{r}
tab  = summarise(data, 
                 n = sum(!is.na(q4)),
                 missing = sum(is.na(q4)),
                 Q1 = round(quantile(q4, 0.25, na.rm=TRUE)),
                 Median = round(quantile(q4, 0.5, na.rm=TRUE)),
                 Q3 = round(quantile(q4, 0.75, na.rm=TRUE)))
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```

The summary statistics above are the median and first to third quartile. `n` is the number answered.

#### Barplot of committee size

```{r}
bplot = ggplot(data = data, aes(x=q4))+
  geom_bar(fill = 'skyblue', col = 'grey88')+
  xlab('Number of committee members')+
  g.theme
bplot
```

Australian HREC committees have a minimum of 8 members.


## Application numbers

The next set of questions examine the number of applications the committee considered in one year (2023). We are only interested in applications that included quantitative data and/or analysis. Our aim is to estimate the total number of applications considered across Australia in one year, hence we also impute the numbers for committees that did not respond to our questionnaire. 

There were two different options for answering the question.

### Option 1: How many applications did the committee consider in 2023 that included quantitative data and/or analysis? Give your best estimate, and a lower and upper range if you are uncertain.

```{r}
# get mean and lower
tab = select(data, id, starts_with('q6_1_')) %>%
  pivot_longer(-id) %>%
  group_by(name) %>%
  summarise( n = sum(!is.na(value)),
                 missing = sum(is.na(value)),
                 Q1 = round(quantile(value, 0.25, na.rm=TRUE)),
                 Median = round(quantile(value, 0.5, na.rm=TRUE)),
                 Q3 = round(quantile(value, 0.75, na.rm=TRUE))) %>%
  mutate(name = nice_rename(name),
         name = str_remove_all(name, ' applications with data and/or analysis')) %>%
  rename('Applications with data and/or analysis' = 'name')
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```

This is the total number of applications with data and/or analysis and the optional lower and upper range.
The summary statistics show the median and first to third quartile.

#### Results where the mean number was outside the lower to upper range

```{r}
outside = filter(data, 
                 q6_1_2 > q6_1_1, # lower greater than mean
                 q6_1_3 < q6_1_1) %>% # upper less than mean
  select(id, q6_1_1, q6_1_2, q6_1_3) %>%
  rename('mean' = 'q6_1_1', 'lower' = 'q6_1_2', 'upper' = 'q6_1_3') %>%
  select('id', 'lower', 'mean', 'upper')
tab = flextable(outside) %>%
  theme_box() %>%
  autofit()
tab
```

The above table is a check that respondents understood the purpose of the lower and upper range. As there are no rows in the table, it appears that respondents understood the question. 

### How many respondents included some uncertainty for option 1

```{r}
# categorise uncertainty as some response to lower or upper interval that is not equal to mean
some_uncertainty = select(data, id, starts_with('q6_1_')) %>%
  filter(!is.na(q6_1_1)) %>% # mean must be there
  mutate(uncertainty = case_when(
    !is.na(q6_1_2)& q6_1_2!=q6_1_1 ~ 'Yes',
    !is.na(q6_1_3)& q6_1_3!=q6_1_1 ~ 'Yes',
    TRUE ~ 'No'
  ))
#
tab =  tabyl(some_uncertainty, uncertainty) %>%
  mutate(percent = round(percent*100)) 
names(tab)[1] = 'Some uncertainty'
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```

### Option 2 part a: How many applications did the committee consider in 2023?

```{r}
tab  = summarise(data, 
                 n = sum(!is.na(q7)),
                 missing = sum(is.na(q7)),
                 Q1 = round(quantile(q7, 0.25, na.rm=TRUE)),
                 Median = round(quantile(q7, 0.5, na.rm=TRUE)),
                 Q3 = round(quantile(q7, 0.75, na.rm=TRUE)))
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```

This is the total number of applications and the optional lower and upper range.

### Option 2 part b: What percentage of these applications included quantitative data and/or analysis? Give your best estimate, and a lower and upper range if you are uncertain.

```{r}
# get mean and lower
tab = select(data, id, starts_with('q8_1_')) %>%
  pivot_longer(-id) %>%
  group_by(name) %>%
  summarise( n = sum(!is.na(value)),
                 missing = sum(is.na(value)),
                 Q1 = round(quantile(value, 0.25, na.rm=TRUE)),
                 Median = round(quantile(value, 0.5, na.rm=TRUE)),
                 Q3 = round(quantile(value, 0.75, na.rm=TRUE))) %>%
  mutate(name = nice_rename(name),
         name = str_remove_all(name, ' percent with data and/or analysis')) %>%
  rename('Percent with data and/or analysis' = 'name')
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```

#### Results where the mean percentage was outside the lower to upper range

```{r}
outside = filter(data, 
                 q8_1_2 > q8_1_1, # lower greater than mean
                 q8_1_3 < q8_1_1) %>% # upper less than mean
  select(id, q8_1_1, q8_1_2, q8_1_3) %>%
  rename('mean' = 'q8_1_1', 'lower' = 'q8_1_2', 'upper' = 'q8_1_3') %>%
  select('id', 'lower', 'mean', 'upper')
tab = flextable(outside) %>%
  theme_box() %>%
  autofit()
tab
```

The above table is a check that respondents understood the purpose of the lower and upper range.

### How many respondents included some uncertainty for option 2

```{r}
# categorise uncertainty as some response to lower or upper interval that is not equal to mean
some_uncertainty = select(data, id, starts_with('q8_1_')) %>%
  filter(!is.na(q8_1_1)) %>% # mean must be there
  mutate(uncertainty = case_when(
    !is.na(q8_1_2)& q8_1_2!=q8_1_1 ~ 'Yes',
    !is.na(q8_1_3)& q8_1_3!=q8_1_1 ~ 'Yes',
    TRUE ~ 'No'
  ))
#
tab =  tabyl(some_uncertainty, uncertainty) %>%
  mutate(percent = round(percent*100)) 
names(tab)[1] = 'Some uncertainty'
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```


### Total number of applications

We combined the results from options 1 and 2 to give an overall estimate of the total number of applications and the uncertainty. We imputed the results for HRECs who did not respond to give a national estimate.

```{r total_imputation}
source('4_total_numbers.R')

# histogram
hplot = ggplot(data = boot_samples, aes(x=total)) + 
  geom_histogram(fill='darkseagreen3', col='grey88')+
  xlab('Total number of applications with quantitative data')+
  ylab('Count')+
  g.theme
hplot
# for text
b_mean = round(mean(boot_samples$total))
b_lower = round(quantile(boot_samples$total, probs = 0.025))
b_upper = round(quantile(boot_samples$total, probs = 0.975))
```

We estimated the number of annual applications with quantitative data and/or analysis using a bootstrap approach using `r n_boot` samples. This allowed us to model the uncertainty in respondents' answers. We used a triangular distribution based on the respondents' lower and upper limits and mean.

The mean number of applications is `r format(b_mean, big.mark=',')` with a 95% confidence interval from `r format(b_lower, big.mark=',')` to `r format(b_upper, big.mark=',')`.

### Does the committee currently have a statistician

```{r}
# make bootstrap confidence intervals that adjust for non-response, see 99_functions
non_response = 187 - 93 # 187 sent email and 93 provided an answer
two_questions = select(data, id, starts_with('q9_'))
tab = bootstrap_intervals(indata = two_questions, non_response = non_response, n_boot = 10000) 
# neaten the table
neat_tab = mutate(tab,
         name = nice_rename(name),
         percent = round(100*r/n),
         lower = round(lower),
         upper = round(upper),
         cell = paste(r, ' (', percent, '%)', sep=''),
         interval = paste(lower, ' to ', upper, sep='')) %>%
    select(name, value, cell, interval) %>%
    rename(
      'Question' = 'name',
      'Response' = 'value',
      'n (%)' = 'cell',
      '95% CI' = 'interval')
ftab = flextable(neat_tab) %>%
  theme_box() %>%
  merge_v(j = 1) %>%
  autofit()
ftab
# export to latex
print(xtable(neat_tab, caption = 'Does the committee currently have a qualified statistician', label='tab:main'), 
      include.rownames=FALSE, 
      math.style.negative=TRUE, 
      table.placement = '!h',
      file='latex/table_have_statistician.tex')
```

The above table shows the frequencies of the two key questions on statisticians. The empty responses are missing.

We used bootstrap confidence intervals to estimate the percentages accounting for non-response using 10,000 bootstrap resamples.

The main outcome is the number and percentage of statisticians who are full committee members.

### Accounting for statistical qualifications

```{r}
# calculate bootstrap intervals
two_questions = select(data, id, starts_with('adequate')) %>%
  mutate(adequate1 = ifelse(adequate1 == 'Unknown', NA, adequate1), # change back to missing for function
         adequate2 = ifelse(adequate2 == 'Unknown', NA, adequate2))
tab = bootstrap_intervals(indata = two_questions, 
                          non_response = non_response, 
                          n_boot = 10000,
                          v1 = 'adequate1',
                          v2 = 'adequate2') 
# neaten the table
neat_tab = mutate(tab,
         name = nice_rename(name),
         percent = round(100*r/n),
         lower = round(lower),
         upper = round(upper),
         cell = paste(r, ' (', percent, '%)', sep=''),
         interval = paste(lower, ' to ', upper, sep='')) %>%
    select(name, value, cell, interval) %>%
    rename(
      'Question' = 'name',
      'Response' = 'value',
      'n (%)' = 'cell',
      '95% CI' = 'interval')
ftab = flextable(neat_tab) %>%
  theme_box() %>%
  merge_v(j = 1) %>%
  autofit()
ftab
```

This table is as per the previous table but after accounting for where the respondents provided the statistician's qualifications. So to be classed as "Yes" the respondents had to respond "Yes" to the question and also provide some statistical qualifications. 

Our secondary outcome is access to statistical expertise (either as a full member or non-member) from a qualified statistician.


### What are the statistician’s qualifications in statistics?

Only for those who answered that they have a statistician on the committee or have access to one.

Respondents could tick multiple boxes.

```{r}
tab = filter(data, q9_1 == 'Yes' | q9_2 == 'Yes') %>% # only those who could answer question
  make_tab_tick_all( var_start = 'q10') %>%
  rename('Role' = 'name') %>%
  mutate(Role = nice_rename(Role))
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```

The percent for "No answer" is for all participants who could have answered the question. The other percents use a denominator of respondents who gave at least one answer.


### Are they currently employed as a statistician?

```{r}
tab =  filter(data, q9_1 == 'Yes' | q9_2 == 'Yes') %>% # only those who answered yes to this previous question
  tabyl(q11) %>%
  mutate(percent = round(percent*100)) %>%
  arrange(desc(n)) %>%
  filter(n > 0) %>%
  rename('Response' = 'q11')
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```

### How do you decide when to consult the statistician? 

Only for respondents who answered that they have access to a statistician.

```{r}
tab =  filter(data, q9_2 == 'Yes') %>% # only those who answered yes to this previous question
  mutate(q12 =  as.character(q12),
         q12 = ifelse(is.na(q12), 'Missing', q12),
         q12 = ifelse(q12=='', 'Missing', q12)) %>% # avoid odd missing values
  tabyl(q12) %>%
  mutate(percent = round(percent*100)) %>%
  arrange(desc(n)) %>%
  filter(n > 0)  %>%
  rename('Response' = 'q12')
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```



### When did your committee last have a statistician as a full member?

Only for respondents who answered that they do not have a statistician on the committee.

```{r}
tab =  filter(data, q9_1 == 'No') %>% # only those who could answer question
  tabyl(q13) %>%
  mutate(percent = round(percent*100)) %>%
  filter(n > 0) %>%
  rename('Response' = 'q13')
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
print(xtable(tab, caption = 'When did your committee last have a statistician as a full member?', label='tab:no_statistician'),       include.rownames = FALSE, 
      table.placement = '!h',
      file='latex/table_when_did.tex')
```


### How are statistical aspects of studies dealt with?

Only for respondents who answered that they do not have a statistician on the committee.

```{r}
tab =  filter(data, q9_1 == 'No') %>% # only those who could answer question
  tabyl(q14) %>%
  mutate(percent = round(percent*100)) %>%
  filter(n > 0) %>%
  rename('Response' = 'q14')
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit() %>%
  width(j=1, width=9.5, unit='cm')
ftab
# output to latex
print(xtable(tab, caption = 'How are statistical aspects of studies dealt with?', label='tab:no_statistician'), 
      include.rownames = FALSE, 
      table.placement = '!h',
      file='latex/table_how_are.tex')
```


### Why don't you have a statistician on the committee? 

Only for respondents who answered that they do not have a statistician on the committee.

Respondents could tick multiple boxes.

```{r why_not}
tab = filter(data, q9_1 == 'No') %>% # only those who could answer question
  make_tab_tick_all(var_start = 'q15') %>%
  rename('Reason' = 'name') %>%
  mutate(Reason = nice_rename(Reason))
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
# output to latex
print(xtable(tab, caption = "Why don`t you have a statistician on the committee?", label='tab:no_statistician'), 
      include.rownames = FALSE, 
      table.placement = '!h',
      file='latex/table_why_not.tex')
```

### Do you consider that the committee needs a statistician as a full member? 

Only for respondents who answered that they do not have a statistician on the committee.

```{r want_full}
tab = filter(data, q9_1 == 'No') %>% # only those who could answer question
  tabyl(q16) %>%
  mutate(percent = round(percent*100)) %>%
  arrange(desc(n)) %>%
  filter(n > 0) %>%
  rename('Response' = 'q16')
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
# output to latex
print(xtable(tab, caption = "Do you consider that the committee needs a statistician as a full member?", label='tab:no_statistician'), 
      include.rownames = FALSE, 
      table.placement = '!h',
      file='latex/table_you_consider.tex')
```


### National Statement wording

Only for respondents who answered that they do not consider that a committee needs a statistician.


```{r national_statement}
tab = filter(data, q16 == 'No') %>% # only those who ...
  tabyl(q18) %>%
  mutate(percent = round(percent*100)) %>%
  arrange(desc(n)) %>%
  filter(n > 0) %>%
  rename('Response' = 'q18')
ftab = flextable(tab) %>%
  theme_box() %>%
  autofit()
ftab
```


## How are statistical concerns handled?

Question wording: ``Consider a hypothetical application. If the statistician advised the committee that the proposed method of data collection and/or analysis were not appropriate to answer the research question(s) then what actions are taken.''

#### The statistician's concerns are discussed amongst the committee

```{r how_handled1}
tab1 = filter(data, q9_1 == 'Yes' | q9_2 == 'Yes') %>% # only those who have a statistician
  tabyl(q21) %>%
  mutate(percent = round(percent*100)) %>%
  filter(n > 0) %>%
  rename('Response' = 'q21')
ftab = flextable(tab1) %>%
  theme_box() %>%
  autofit()
ftab
```

#### The application is sent back to the researchers and they are asked to address the statistician's concerns

```{r how_handled2}
tab2 = filter(data, q9_1 == 'Yes' | q9_2 == 'Yes') %>% # only those who have a statistician
  tabyl(q22) %>%
  mutate(percent = round(percent*100)) %>%
  filter(n > 0) %>%
  rename('Response' = 'q22')
ftab = flextable(tab2) %>%
  theme_box() %>%
  autofit()
ftab
```


#### Summary plot

```{r, fig.width=7}
# combine tables and fix missing
to_plot = bind_rows(tab1, tab2, .id = 'question') %>%
  mutate(question = case_when(
    question == 1 ~ "The statistician’s concerns are discussed amongst the committee",
    question == 2 ~ "The application is sent back to the researchers and they are asked to address the statistician's concerns"
  )) %>%
  mutate(Response = ifelse(is.na(Response), 'Missing', Response))
# labels and colours
xlabels =  c("Always","Usually","Sometimes","Rarely","Never","Missing")
colours = c(RColorBrewer::brewer.pal(name='Blues', n=5), 'grey77')
#
bplot = ggplot(data = to_plot, aes(fill = Response, x = question, y = n))+
  geom_bar(stat = 'identity')+
  scale_x_discrete(expand=c(0.01,0.01), labels = function(x) lapply(strwrap(x, width = 20, simplify = FALSE), paste, collapse="\n"))+
  scale_y_continuous(expand=c(0,NA))+ # reduce gaps to axis
  xlab('')+
  ylab('Frequency')+
  coord_flip()+
  scale_fill_manual(name = NULL, values = colours, labels=xlabels)+
  g.theme
bplot
# make latex table for paper
for_latex = full_join(tab1, tab2, by='Response') %>%
  mutate(cell1 = paste(n.x, ' (', percent.x, '%)', sep=''), # make percent and number cell
         cell2 = paste(n.y, ' (', percent.y, '%)', sep='')) %>%
  select('Response', 'cell1', 'cell2')
# export to latex - transposed
print(xtable(t(for_latex), caption = 'Summary table of how statistical issues are dealt with', label='tab:with_statistician'), 
      include.rownames=FALSE, 
      math.style.negative=TRUE, 
      table.placement = '!h',
      file='latex/table_statistical_issues.tex')
```


## Committees with access to a statistician

Here we estimate the number and percentage of committees that have access to a qualified statistician. 
This is those committees with a statistician as a member, or where there is a statistician as a non-member who can be consulted. For both cases, we verified that the statistician had qualifications in statistics. The unknown response is where these questions were not answered.

```{r}
tab = tabyl(data, any_adequate) %>%
  mutate(percent = round(percent*100))
names(tab)[1] = 'Access to statistician'
ftab = flextable(tab) %>%
  theme_box() %>%
  colformat_double(j=3, digits=0) %>%
  autofit()
ftab
```

### Estimated national application numbers by access to statistician

Here we stratify the estimated total number of applications according to whether the committee has access to a statistician. We only use applications that included quantitative data. We include the uncertainty due to: the ranges provided by the respondents, the results from committees that did not respond. The estimates are national annual totals and are the mean and a 90% bootstrap interval.

```{r}
# boot_samples_by data is from 4_total_numbers.R
# estimate mean and bootstrap interval
stats = group_by(boot_samples_by, any_adequate) %>%
  summarise(mean = round(mean(total)),
            lower = round(quantile(total, 0.05)), # 90% CI
            upper = round(quantile(total, 0.95))) %>%
  ungroup()

#
bplot = ggplot(data = stats, aes(x=factor(any_adequate), label = mean, y=mean, ymin=lower, ymax=upper, col=factor(any_adequate)))+
  xlab('')+
  ylab('National application numbers')+
  geom_point(size=5)+
  geom_errorbar(width=0, linewidth=1.2)+
  g.theme+
  coord_flip()+
  theme(legend.position = 'none')+
  scale_color_manual('Access to qualified\nstatistician', values = c("indianred4", "green3", "orange3"))
bplot
# export
ggsave(bplot, filename='figures/4_numbers.jpg', width=5, height=3, units='in', dpi=500)

# alternative version for slide with numbers at mean
bplotl = bplot + 
  theme(axis.text = element_text(size=16), axis.title = element_text(size=14))+ # make text bigger
  geom_text(vjust=-1) # add number labels
# export
ggsave(bplotl, filename='figures/4_numbers_plus.jpg', width=5.5, height=3, units='in', dpi=500)
```

##### Table of the plotted data

```{r}
ftab = mutate(stats, 
               mean = round(mean),
               lower = round(lower),
               upper = round(upper)) %>%
  flextable() %>%
  autofit()
ftab
```

Results show the mean and lower and upper limit from the bootstrap confidence interval.

# Item-missing data

The table below show item-missing data with panels that match the questionnaire logic. This avoids data being wrongly flagged as missing when the respondent never saw that question. 

```{r item_missing}
source('4_missing_table.R')
# make single cell
missing_table = mutate(missing_table,
                       cell = paste(missing, ' (', percent, ')', sep='')) %>%
  select(labels, n, cell) %>%
  rename('Question' = 'labels',
         'Max response' = 'n',
         'Missing, n (%)' = 'cell')

#
ftab = flextable(missing_table) %>%
  theme_box() %>%
  width(j=1, width = 12, unit = 'cm')
ftab
# output to latex
print(xtable(missing_table, caption = 'Number and percentage of item-missing questions', label='tab:item_missing'), 
      include.rownames = FALSE, 
      table.placement = '!h',
      file='latex/table_item_missing.tex')
```
